+++ Bud.Core/src/main/cs/Properties/AssemblyInfo.cs
@@ -1,11 +1,11 @@
 using System.Reflection;
 
 [assembly: AssemblyTitle("Bud.Core")]
-[assembly: AssemblyDescription("0.1.3-dev")]
+[assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("")]
+[assembly: AssemblyProduct("Bud.Core v0.2.1-dev")]
 [assembly: AssemblyCopyright("Matej Urbas")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
-[assembly: AssemblyVersion("0.1.3")]
+[assembly: AssemblyVersion("0.2.1")]